monitoring/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
psutil==7.2.2
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
wakeonlan==3.1.0
